Upgrade JsonReplace to ReplaceItem
