--- packages/core/src/computers/JsonReplace/index.ts
@@ -1,1 +1,0 @@
-export { JsonReplace } from './JsonReplace'
