+++ packages/core/src/computers/index.ts
@@ -10,7 +10,7 @@ export { Filter } from './Filter';
 export { Ignore } from './Ignore'
 export { InstantThrow } from './InstantThrow';
 export { Input } from './Input';
-export { JsonReplace } from './JsonReplace'
+export { ReplaceItem } from './ReplaceItem'
 export { Log } from './Log';
 export { Map } from './Map'
 // export { Merge } from './Merge';
